+++ tests/src/containers/freelist_tests.c
@@ -13,7 +13,7 @@ u8 freelist_should_create_and_destroy() {
     u8 failed = false;
 
     freelist list;
-    u32 total_size = 1024; // Managing 1KB
+    u64 total_size = 1024; // Managing 1KB
     u64 memory_requirement = 0;
 
     // 1. Get memory requirement
@@ -40,15 +40,15 @@ u8 freelist_should_allocate_successfully() {
     u8 failed = false;
 
     freelist list;
-    u32 total_size = 512;
+    u64 total_size = 512;
     u64 memory_requirement = 0;
 
     freelist_create(total_size, &memory_requirement, 0, 0);
     void *memory = kallocate(memory_requirement, MEMORY_TAG_ARRAY);
     freelist_create(total_size, &memory_requirement, memory, &list);
 
-    u32 offset = 0;
-    u32 alloc_size = 64;
+    u64 offset = 0;
+    u64 alloc_size = 64;
 
     // Allocate 64 bytes
     b8 result = freelist_allocate_block(&list, alloc_size, &offset);
@@ -68,15 +68,15 @@ u8 freelist_should_allocate_and_free_successfully() {
     u8 failed = false;
 
     freelist list;
-    u32 total_size = 512;
+    u64 total_size = 512;
     u64 memory_requirement = 0;
 
     freelist_create(total_size, &memory_requirement, 0, 0);
     void *memory = kallocate(memory_requirement, MEMORY_TAG_ARRAY);
     freelist_create(total_size, &memory_requirement, memory, &list);
 
-    u32 offset = 0;
-    u32 alloc_size = 128;
+    u64 offset = 0;
+    u64 alloc_size = 128;
 
     // Alloc
     b8 result = freelist_allocate_block(&list, alloc_size, &offset);
@@ -100,14 +100,14 @@ u8 freelist_should_handle_multiple_allocations() {
     u8 failed = false;
 
     freelist list;
-    u32 total_size = 1024;
+    u64 total_size = 1024;
     u64 memory_requirement = 0;
 
     freelist_create(total_size, &memory_requirement, 0, 0);
     void *memory = kallocate(memory_requirement, MEMORY_TAG_ARRAY);
     freelist_create(total_size, &memory_requirement, memory, &list);
 
-    u32 offset1 = 0, offset2 = 0, offset3 = 0;
+    u64 offset1 = 0, offset2 = 0, offset3 = 0;
 
     // Alloc 1
     expect_to_be_true(freelist_allocate_block(&list, 100, &offset1));
@@ -135,14 +135,14 @@ u8 freelist_should_fail_oversized_allocation() {
     u8 failed = false;
 
     freelist list;
-    u32 total_size = 100;
+    u64 total_size = 100;
     u64 memory_requirement = 0;
 
     freelist_create(total_size, &memory_requirement, 0, 0);
     void *memory = kallocate(memory_requirement, MEMORY_TAG_ARRAY);
     freelist_create(total_size, &memory_requirement, memory, &list);
 
-    u32 offset = 0;
+    u64 offset = 0;
 
     // Try to allocate 101 bytes from 100 byte list
     b8 result = freelist_allocate_block(&list, 101, &offset);
@@ -160,14 +160,14 @@ u8 freelist_should_clear_successfully() {
     u8 failed = false;
 
     freelist list;
-    u32 total_size = 512;
+    u64 total_size = 512;
     u64 memory_requirement = 0;
 
     freelist_create(total_size, &memory_requirement, 0, 0);
     void *memory = kallocate(memory_requirement, MEMORY_TAG_ARRAY);
     freelist_create(total_size, &memory_requirement, memory, &list);
 
-    u32 offset = 0;
+    u64 offset = 0;
     freelist_allocate_block(&list, 200, &offset);
 
     // Check space is reduced
@@ -192,14 +192,14 @@ u8 freelist_should_reuse_freed_space() {
     u8 failed = false;
 
     freelist list;
-    u32 total_size = 1024;
+    u64 total_size = 1024;
     u64 memory_requirement = 0;
 
     freelist_create(total_size, &memory_requirement, 0, 0);
     void *memory = kallocate(memory_requirement, MEMORY_TAG_ARRAY);
     freelist_create(total_size, &memory_requirement, memory, &list);
 
-    u32 offset1 = 0, offset2 = 0, offset3 = 0;
+    u64 offset1 = 0, offset2 = 0, offset3 = 0;
 
     // Allocate 3 blocks
     freelist_allocate_block(&list, 100, &offset1);
@@ -209,7 +209,7 @@ u8 freelist_should_reuse_freed_space() {
     // Free the middle block
     freelist_free_block(&list, 100, offset2);
 
-    u32 offset_new = 0;
+    u64 offset_new = 0;
     // Allocate a block that fits in the hole
     b8 result = freelist_allocate_block(&list, 100, &offset_new);
